adt/kv: Fix regression
Tracking a memory corruption of private code ended up copying the
original djb code here without considering the model here (key->value).
The walk() is guaranteed to return a "leaf" (considering memory to be
correctly aligned).

A manual guarantee was added just to be sure (as setalloc is officially
part of the API).